fix: accept float AI scores on GET /properties (BIN-56) (#56)
* fix: accept float AI scores on GET /properties (BIN-56)

PropertyModel typed condition_score/sentiment_score as int while AI
enrichment stores floats in [0.0, 1.0], causing ResponseValidationError
500s. Align the schema and lock it with unit/integration/contract tests;
migrate the DB in CI contract so 500s are no longer silently ignored.

* test: bypass Redis rate limit in properties schema unit test

Unit CI has no Redis; slowapi was ConnectionError'ing before
response_model validation could run.

+++ src/api/schemas.py
@@ -47,8 +47,9 @@ class PropertyModel(BaseModel):
     ai_issues: List[str] = []
     ai_green_flags: List[str] = []
     ai_red_flags: List[str] = []
-    condition_score: Optional[int] = None
-    sentiment_score: Optional[int] = None
+    # AI domain scores are floats in [0.0, 1.0] (see VisualResult / SentimentResult).
+    condition_score: Optional[float] = None
+    sentiment_score: Optional[float] = None
     stat_category: Optional[str] = None
     stat_reasoning: Optional[str] = None
     deal_summary: Optional[str] = None
